Feat: Add category and purchase products operations

+++ services/product-service/src/main/java/com/riquelmemr/productservice/data/ProductData.java
@@ -14,8 +14,8 @@ public class ProductData {
     private Long id;
     private String name;
     private String description;
-    private double stock;
-    private double reserved;
+    private int stock;
+    private int reserved;
     private BigDecimal price;
     private Set<CategoryModel> categories;
