+++ services/product-service/src/main/java/com/riquelmemr/productservice/dto/ProductResponse.java
@@ -4,12 +4,19 @@ import java.math.BigDecimal;
 import java.util.Set;
 
 public record ProductResponse(
+
         Long id,
+
         String name,
+
         String description,
-        double stock,
-        double reserved,
+
+        int stock,
+
+        int reserved,
+
         BigDecimal price,
+
         Set<CategoryResponse> categories
 ) {
 }
