+++ services/product-service/src/main/resources/db/migration/V1__create_product_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE product.tb_product (
     id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    stock DOUBLE PRECISION NOT NULL DEFAULT 0,
-    reserved DOUBLE PRECISION NOT NULL DEFAULT 0,
+    stock INTEGER NOT NULL DEFAULT 0,
+    reserved INTEGER NOT NULL DEFAULT 0,
     price NUMERIC(19,2) NOT NULL
 );
